fix code r/t review
