+++ build.rs
@@ -41,7 +41,8 @@ const DATA: &'static [(&'static str, &'static [&'static str], &'static [&'static
     ("enclaveapi", &["basetsd", "minwinbase", "minwindef", "ntdef", "winnt"], &["kernel32"]),
     ("evntprov", &["basetsd", "guiddef", "minwindef", "winnt"], &["advapi32"]),
     ("evntrace", &["basetsd", "evntcons", "evntprov", "guiddef", "handleapi", "minwindef", "timezoneapi", "vadefs", "winnt", "wmistr"], &["advapi32"]),
-    ("fltuser", &["minwinbase", "minwindef", "ntdef", "winnt"], &["fltlib"]),
+    ("fltuser", &["fltuserstructures", "minwinbase", "minwindef"], &["fltlib"]),
+    ("fltuserstructures", &["ntdef", "winnt"], &[]),
     ("guiddef", &[], &[]),
     ("hidclass", &["guiddef", "minwindef", "winioctl", "winnt"], &[]),
     ("hidpi", &["hidusage", "minwindef", "ntdef", "ntstatus", "winnt"], &["hid"]),
