+++ tests/structs_x86_64.rs
@@ -4272,23 +4272,23 @@ fn um_fileapi() {
 }
 #[cfg(feature = "fltuser")] #[test]
 fn um_fltuser() {
-    use winapi::um::fltuser::*;
+    use winapi::shared::fltuserstructures::*;
     assert_eq!(size_of::<FILTER_FULL_INFORMATION>(), 16);
     assert_eq!(align_of::<FILTER_FULL_INFORMATION>(), 4);
-    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_u_s_MiniFilter>(), 16);
-    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_u_s_MiniFilter>(), 4);
-    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_u_s_LegacyFilter>(), 4);
-    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_u_s_LegacyFilter>(), 2);
-    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_u>(), 16);
-    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_u>(), 4);
+    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_MiniFilter>(), 16);
+    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_MiniFilter>(), 4);
+    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_LegacyFilter>(), 4);
+    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type_LegacyFilter>(), 2);
+    assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type>(), 16);
+    assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION_Type>(), 4);
     assert_eq!(size_of::<FILTER_AGGREGATE_BASIC_INFORMATION>(), 24);
     assert_eq!(align_of::<FILTER_AGGREGATE_BASIC_INFORMATION>(), 4);
-    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_u_s_MiniFilter>(), 20);
-    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_u_s_MiniFilter>(), 4);
-    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_u_s_LegacyFilter>(), 12);
-    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_u_s_LegacyFilter>(), 4);
-    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_u>(), 20);
-    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_u>(), 4);
+    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_MiniFilter>(), 20);
+    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_MiniFilter>(), 4);
+    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_LegacyFilter>(), 12);
+    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type_LegacyFilter>(), 4);
+    assert_eq!(size_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type>(), 20);
+    assert_eq!(align_of::<FILTER_AGGREGATE_STANDARD_INFORMATION_Type>(), 4);
     assert_eq!(size_of::<FILTER_VOLUME_BASIC_INFORMATION>(), 4);
     assert_eq!(align_of::<FILTER_VOLUME_BASIC_INFORMATION>(), 2);
     assert_eq!(size_of::<FILTER_VOLUME_STANDARD_INFORMATION>(), 20);
@@ -4299,12 +4299,12 @@ fn um_fltuser() {
     assert_eq!(align_of::<INSTANCE_PARTIAL_INFORMATION>(), 4);
     assert_eq!(size_of::<INSTANCE_FULL_INFORMATION>(), 20);
     assert_eq!(align_of::<INSTANCE_FULL_INFORMATION>(), 4);
-    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_u_s_MiniFilter>(), 32);
-    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_u_s_MiniFilter>(), 4);  
-    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_u_s_LegacyFilter>(), 20);
-    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_u_s_LegacyFilter>(), 4);
-    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_u>(), 32);
-    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_u>(), 4);
+    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_MiniFilter>(), 32);
+    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_MiniFilter>(), 4);
+    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_LegacyFilter>(), 20);
+    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type_LegacyFilter>(), 4);
+    assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type>(), 32);
+    assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION_Type>(), 4);
     assert_eq!(size_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION>(), 40);
     assert_eq!(align_of::<INSTANCE_AGGREGATE_STANDARD_INFORMATION>(), 4);
     assert_eq!(size_of::<FILTER_MESSAGE_HEADER>(), 16);
